2c: Split main.go into main.go + wiring.go
